% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{add_hooks}
\alias{add_hooks}
\title{Add hooks to the installed package}
\usage{
add_hooks(pkg_name, lib, fix_mcexit = FALSE)
}
\arguments{
\item{pkg_name}{name of the package to add hooks to}

\item{lib}{the library path to look in}

\item{fix_mcexit}{whether to add the fix for mcparallel:::mcexit}
}
\description{
Installed packages have lazy loading code to setup the lazy load database at
pkg_name/R/pkg_name. This function adds a user level onLoad Hook to the
package which calls \code{covr::trace_environment}, so the package environment is
traced when the package is loaded.
}
\details{
It also adds a finalizer that saves the tracing information to the package
namespace environment which is run when the ns is garbage collected or the
process ends. This ensures the tracing count information will be written
regardless of how the process terminates.
}
