% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{as_coverage}
\alias{as_coverage}
\title{Convert a counters object to a coverage object}
\usage{
as_coverage(counters = NULL, ...)
}
\arguments{
\item{counters}{An environment of covr trace results to convert to a coverage
object. If \code{counters} is not provided, the \code{covr} namespace value
\code{.counters} is used.}

\item{...}{Additional attributes to include with the coverage object.}
}
\description{
Convert a counters object to a coverage object
}
