% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{as_coverage_with_tests}
\alias{as_coverage_with_tests}
\title{Clean and restructure counter tests for a coverage object}
\usage{
as_coverage_with_tests(counters)
}
\arguments{
\item{counters}{An environment of covr trace results to convert to a coverage
object. If \code{counters} is not provided, the \code{covr} namespace value
\code{.counters} is used.}
}
\description{
For tests produced with \code{options(covr.record_tests)}, prune any unused
records in the $tests$tally matrices of each trace and get rid of the
wrapping $tests environment (reassigning with value of $tests$tally)
}
