% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure}
\alias{azure}
\title{Run covr on a package and output the result so it is available on Azure Pipelines}
\usage{
azure(..., coverage = package_coverage(..., quiet = quiet),
  filename = "coverage.xml", quiet = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{\link[=package_coverage]{package_coverage()}}}

\item{coverage}{an existing coverage object to submit, if \code{NULL},
\code{\link[=package_coverage]{package_coverage()}} will be called with the arguments from
\code{...}}

\item{filename}{the name of the Cobertura XML file}

\item{quiet}{if \code{FALSE}, print the coverage before submission.}
}
\description{
Run covr on a package and output the result so it is available on Azure Pipelines
}
