% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobertura.R
\name{cobertura_dtd}
\alias{cobertura_dtd}
\title{Generate cobertura DTD}
\usage{
cobertura_dtd()
}
\value{
String that contains the cobertura DTD
}
\description{
Generate cobertura DTD
}
\author{
Willem Ligtenberg
}

