% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{count_branch}
\alias{count_branch}
\title{increment a given branch counter (used for implicit branches)}
\usage{
count_branch(key)
}
\arguments{
\item{key}{generated with \code{\link[=key]{key()}}}
}
\description{
increment a given branch counter (used for implicit branches)
}
\keyword{internal}
