% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{count_test}
\alias{count_test}
\title{append a testref to a counter}
\usage{
count_test(key)
}
\arguments{
\item{key}{generated with \code{\link[=key]{key()}}}
}
\description{
append a testref to a counter
}
\keyword{internal}
