% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{coverage_to_list}
\alias{coverage_to_list}
\title{Convert a coverage dataset to a list}
\usage{
coverage_to_list(x = package_coverage())
}
\arguments{
\item{x}{a coverage dataset, defaults to running \code{package_coverage()}.}
}
\value{
A list containing coverage result for each individual file and the whole package
}
\description{
Convert a coverage dataset to a list
}
