% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{shine}
\alias{shine}
\title{Deprecated Functions
These functions are Deprecated in this release of covr, they will be
marked as Defunct and removed in a future version.}
\usage{
shine(...)
}
\description{
Deprecated Functions
These functions are Deprecated in this release of covr, they will be
marked as Defunct and removed in a future version.
}
\keyword{internal}
