% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{current_test_call_count}
\alias{current_test_call_count}
\title{Retrieve the number of times the test call was called}
\usage{
current_test_call_count()
}
\value{
An integer value representing the number of calls of the current
call into the package from the testing suite.
}
\description{
A single test expression might be evaluated many times. Each time the same
expression is called, the call count is incremented.
}
