% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{current_test_index}
\alias{current_test_index}
\title{Retrieve the index for the test in \code{.counters$tests}}
\usage{
current_test_index()
}
\value{
An integer index for the test call
}
\description{
If the test was encountered before, the index will be the index of the test
in the logged tests list. Otherwise, the index will be the next index beyond
the length of the tests list.
}
\keyword{internal}
