% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{current_test_key}
\alias{current_test_key}
\title{Build key for the current test}
\usage{
current_test_key()
}
\value{
A unique character string if the test call has a srcref, or an empty
string otherwise.
}
\description{
If the current test has a srcref, a unique character key is built from its
srcref. Otherwise, an empty string is returned.
}
\keyword{internal}
