% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display_name.R
\name{display_name<-}
\alias{display_name<-}
\title{Set the display name for an object}
\usage{
display_name(x) <- value
}
\arguments{
\item{x}{object to set the display name}

\item{value}{what to set the display name to}
}
\description{
Set the display name for an object
}

