% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_name.R
\name{display_name}
\alias{display_name}
\title{Retrieve the path name (filename) for each coverage object}
\usage{
display_name(x)
}
\arguments{
\item{x}{A coverage object}
}
\description{
Retrieve the path name (filename) for each coverage object
}
\keyword{internal}
