% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{environment_coverage}
\alias{environment_coverage}
\title{Calculate coverage of an environment}
\usage{
environment_coverage(env = parent.frame(), test_files,
  line_exclusions = NULL, function_exclusions = NULL)
}
\arguments{
\item{env}{The environment to be instrumented.}

\item{line_exclusions}{a named list of files with the lines to exclude from
each file.}

\item{function_exclusions}{a vector of regular expressions matching function
names to exclude. Example \code{print\\\.} to match print methods.}

\item{test_code}{A character vector of test code}
}
\description{
Calculate coverage of an environment
}
