% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/covr.R
\name{environment_coverage_}
\alias{environment_coverage_}
\title{calculate the coverage on an environment after evaluating some expressions.}
\usage{
environment_coverage_(env, exprs, enc = parent.frame())
}
\arguments{
\item{env}{the environment to take function definitions from}

\item{exprs}{a list of parsed expressions to be evaluated.}

\item{enc}{The enclosing environment from which the expressions should be
evaluated}
}
\description{
This function does not use non_standard evaluation so is more appropriate
for use in other functions.
}

