% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{file_report}
\alias{file_report}
\title{A coverage report for a specific file}
\usage{
file_report(x = package_coverage(), file = NULL,
  out_file = file.path(tempdir(), paste0(get_package_name(x),
  "-file-report.html")), browse = interactive())
}
\arguments{
\item{x}{a coverage dataset, defaults to running \code{package_coverage()}.}

\item{file}{The file to report on, if \code{NULL}, use the first file in the
coverage output.}

\item{out_file}{The output file}

\item{browse}{whether to open a browser to view the report.}
}
\description{
A coverage report for a specific file
}
