% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/covr.R
\name{function_coverage}
\alias{function_coverage}
\title{Calculate test coverage for a package}
\usage{
function_coverage(names, ..., enc = parent.frame())
}
\arguments{
\item{names}{function names.}

\item{...}{expressions to run.}

\item{enc}{the enclosing environment which to run the expressions.}
}
\description{
Calculate test coverage for a package
}

