% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{has_srcref}
\alias{has_srcref}
\title{Is the source bound to the expression}
\usage{
has_srcref(expr)
}
\arguments{
\item{expr}{A language object which may have a \code{srcref} attribute}
}
\value{
A logical value indicating whether the language object has source
}
\description{
Is the source bound to the expression
}
