% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{in_covr}
\alias{in_covr}
\title{Determine if code is being run in covr}
\usage{
in_covr()
}
\description{
covr functions set the environment variable \code{R_COVR} when they are running.
\code{\link[=in_covr]{in_covr()}} returns \code{TRUE} if this environment variable is set and \code{FALSE}
otherwise.
}
\examples{
if (require(testthat)) {
  testthat::skip_if(in_covr())
}
}
