% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{is_covr_count_call}
\alias{is_covr_count_call}
\title{Is the expression a call to covr:::count}
\usage{
is_covr_count_call(expr)
}
\arguments{
\item{expr}{A language object}
}
\value{
A logical value indicating whether the object is a call to
\code{covr:::count}.
}
\description{
Is the expression a call to covr:::count
}
