% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{new_branch}
\alias{new_branch}
\title{initialize a new branch}
\usage{
new_branch(srcref, parent_functions, parent_srcref, default)
}
\arguments{
\item{parent_functions}{the functions that this srcref is contained in.}

\item{src_ref}{a \code{\link[base:srcref]{base::srcref()}}}
}
\description{
initialize a new branch
}
\keyword{internal}
