% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{new_counter}
\alias{new_counter}
\title{initialize a new counter}
\usage{
new_counter(src_ref)
}
\arguments{
\item{src_ref}{a \code{\link[base]{srcref}}}
}
\description{
initialize a new counter
}

