% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/covr.R
\name{package_coverage}
\alias{package_coverage}
\title{Calculate test coverage for a package}
\usage{
package_coverage(path = ".", ..., relative_path = TRUE)
}
\arguments{
\item{path}{file path to the package}

\item{...}{expressions to run}

\item{relative_path}{whether to output the paths as relative or absolute
paths.}
}
\description{
Calculate test coverage for a package
}

