% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{percent_coverage}
\alias{percent_coverage}
\title{Provide percent coverage of package}
\usage{
percent_coverage(coverage_result)
}
\arguments{
\item{coverage_result}{the coverage object returned from \code{\link{package_coverage}}}
}
\description{
Provide percent coverage of package
}

