% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.branch_coverage}
\alias{print.branch_coverage}
\title{Print the branch coverage of a coverage object}
\usage{
\method{print}{branch_coverage}(x, group = c("filename", "functions"), ...)
}
\arguments{
\item{x}{the branch coverage object to be printed}

\item{group}{whether to group coverage by filename or function}

\item{...}{additional arguments ignored}
}
\value{
The branch coverage object (invisibly).
}
\description{
Print the branch coverage of a coverage object
}
