% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.coverage}
\alias{print.coverage}
\title{Print a coverage object}
\usage{
\method{print}{coverage}(
  x,
  group = c("filename", "functions"),
  by = "line",
  include_branches = FALSE,
  ...
)
}
\arguments{
\item{x}{the coverage object to be printed}

\item{group}{whether to group coverage by filename or function}

\item{by}{whether to count coverage by line or expression}

\item{include_branches}{whether to include branch coverage or not}

\item{...}{additional arguments ignored}
}
\value{
The coverage object (invisibly).
}
\description{
Print a coverage object
}
