% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.R
\name{R}
\alias{R}
\title{Run R xxx from within R}
\usage{
R(options, path = tempdir(), env_vars = NULL, ...)
}
\arguments{
\item{options}{a character vector of options to pass to the command}

\item{path}{the directory to run the command in.}

\item{env_vars}{environment variables to set before running the command.}

\item{...}{additional arguments passed to \code{\link{system_check}}}
}
\description{
Run R xxx from within R
}

