% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{record_test}
\alias{record_test}
\title{set the current test}
\usage{
record_test(key)
}
\arguments{
\item{key}{generated with \code{\link[=key]{key()}}}
}
\description{
set the current test
}
\keyword{internal}
