% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{scrub_anon_fn_envs}
\alias{scrub_anon_fn_envs}
\title{Scrub Environments from Anonymous Function Calls}
\usage{
scrub_anon_fn_envs(x)
}
\arguments{
\item{calls}{A call object which we should}
}
\description{
When a sys call is to an anonymous function, the anonymous function is
included as the object that the call is against. This function binds its
calling environment, which may bloat the sys calls object and in worst
case scenario might contain C API memory pointers that can produce
segmentation faults when serialized to disk.
}
\details{
As the environment data is not useful after saving the coverage traces to
disk and restoring them in the parent R session, these are discarded and
replaced with an \code{emptyenv()}
}
