% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subprocess.R
\name{subprocess}
\alias{subprocess}
\title{Run R commands in a R subprocess}
\usage{
subprocess(code, calling_env = parent.frame(), global_env = .GlobalEnv,
  quiet = F, clean = TRUE)
}
\arguments{
\item{code}{expression to call in the subprocess}

\item{calling_env}{the environment of the calling function.}

\item{global_env}{the global environment to load in the subprocess.}

\item{quiet}{whether to echo the R command run.}

\item{clean}{whether to clean the Rout files generated from the subprocess.}
}
\description{
this function runs a R subprocess then given \code{code}, trying to
replicate as much of the calling environment as possible, then returning the
results to the calling environment.
}

