% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{tally_branch_coverage}
\alias{tally_branch_coverage}
\title{Tally branch coverage}
\usage{
tally_branch_coverage(x)
}
\arguments{
\item{x}{the branch coverage object to be printed}
}
\value{
a \code{data.frame} of branch coverage
}
\description{
Tally branch coverage
}
