% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{tally_coverage}
\alias{tally_coverage}
\title{Tally coverage by line or expression}
\usage{
tally_coverage(x, by = c("line", "expression"))
}
\arguments{
\item{x}{the coverage object returned from \code{\link{package_coverage}}}

\item{by}{whether to tally coverage by line or expression}
}
\value{
a \code{data.frame} of coverage tallied by line or expression.
}
\description{
Tally coverage by line or expression
}
