% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobertura.R
\name{to_cobertura}
\alias{to_cobertura}
\title{Create a Cobertura XML file}
\usage{
to_cobertura(cov, filename = "cobertura.xml")
}
\arguments{
\item{cov}{the coverage object returned from \code{\link{package_coverage}}}

\item{filename}{the name of the Cobertura XML file}
}
\description{
Create a Cobertura XML file
}
\author{
Willem Ligtenberg
}

