% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcov.R
\name{to_lcov}
\alias{to_lcov}
\title{Create an LCOV file}
\usage{
to_lcov(cov, filename = "coverage.lcov")
}
\arguments{
\item{cov}{the coverage object returned from \code{\link[=package_coverage]{package_coverage()}} or \code{\link[=file_coverage]{file_coverage()}}}

\item{filename}{the name of the LCOV file to write}
}
\description{
Create an LCOV-compliant text report.
The format is documented at:
https://github.com/linux-test-project/lcov/
}
