% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonarqube.R
\name{to_sonarqube}
\alias{to_sonarqube}
\title{Create a SonarQube Generic XML file for test coverage according to
https://docs.sonarqube.org/latest/analysis/generic-test/
Based on cobertura.R}
\usage{
to_sonarqube(cov, filename = "sonarqube.xml")
}
\arguments{
\item{cov}{the coverage object returned from \code{\link[=package_coverage]{package_coverage()}}}

\item{filename}{the name of the SonarQube Generic XML file}
}
\description{
This functionality requires the xml2 package be installed.
}
\author{
Talkdesk Inc.
}
