% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{trace_calls}
\alias{trace_calls}
\title{trace each call with a srcref attribute}
\usage{
trace_calls(x, srcref = NULL)
}
\arguments{
\item{x}{the call}

\item{srcref}{argument used to set the srcref of the current call when recursing}
}
\value{
a modified expression with count calls inserted before each previous
call.
}
\description{
This function calls itself recursively so it can properly traverse the AST.
}
\seealso{
\url{http://adv-r.had.co.nz/Expressions.html}
}

