% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{truncate_call}
\alias{truncate_call}
\title{Truncate call objects to limit the number of arguments}
\usage{
truncate_call(call_obj, limit = 10000)
}
\arguments{
\item{call_obj}{A (possibly large) \code{call} object}

\item{limit}{A \code{call} length limit to impose}
}
\value{
The \code{call_obj} with arguments trimmed
}
\description{
A helper to circumvent R errors when deserializing large call objects from
Rds. Trims the number of arguments in a call object, and replaces the last
argument with a \verb{<truncated>} symbol.
}
