% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{update_current_test}
\alias{update_current_test}
\title{Update current test if unit test expression has progressed}
\usage{
update_current_test(key)
}
\arguments{
\item{key}{generated with \code{\link[=key]{key()}}}
}
\description{
Update current test if unit test expression has progressed
}
\keyword{internal}
