% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{Retrieve the value from an object}
\usage{
value(x, ...)
}
\arguments{
\item{x}{object from which to retrieve the value}

\item{...}{additional arguments passed to methods}
}
\description{
Retrieve the value from an object
}

