% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{zero_coverage}
\alias{zero_coverage}
\title{Provide locations of zero coverage}
\usage{
zero_coverage(coverage_result)
}
\arguments{
\item{coverage_result}{a coverage object returned from \code{\link{package_coverage}}}
}
\description{
When examining the test coverage of a package, it is useful to know if there are
any locations where there is \bold{0} test coverage.
}

