% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/codecov.R
\name{codecov}
\alias{codecov}
\title{Run covr on a package and upload the result to codecov.io}
\usage{
codecov(path = ".", ..., base_url = "https://codecov.io")
}
\arguments{
\item{path}{file path to the package}

\item{...}{additional arguments passed to \code{\link{package_coverage}}}

\item{base_url}{Codecov url (change for Enterprise)}
}
\description{
Run covr on a package and upload the result to codecov.io
}

