% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{count_test}
\alias{count_test}
\title{Append a test trace to a counter, updating global current test}
\usage{
count_test(key)
}
\arguments{
\item{key}{generated with \code{\link[=key]{key()}}}
}
\description{
Append a test trace to a counter, updating global current test
}
\keyword{internal}
