% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coveralls.R
\name{coveralls}
\alias{coveralls}
\title{Run covr on a package and upload the result to coveralls}
\usage{
coveralls(path = ".", ...)
}
\arguments{
\item{path}{file path to the package}

\item{...}{additional arguments passed to \code{\link{package_coverage}}}
}
\description{
Run covr on a package and upload the result to coveralls
}

