% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\docType{package}
\name{covr-package}
\alias{covr}
\alias{covr-package}
\title{covr: Test coverage for packages}
\description{
covr tracks and reports code coverage for your package and (optionally)
upload the results to a coverage service like 'Codecov' \url{http://codecov.io} or
'Coveralls' \url{http://coveralls.io}. Code coverage is a measure of the amount of
code being exercised by a set of tests. It is an indirect measure of test
quality and completeness. This package is compatible with any testing
methodology or framework and tracks coverage of both R code and compiled
C/C++/FORTRAN code.
}
\details{
A coverage report can be used to inspect coverage for each line in your
package. Using \code{report()} requires the additional dependencies \code{DT} and \code{htmltools}.\preformatted{# If run with no arguments `report()` implicitly calls `package_coverage()`
report()
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/covr}
  \item Report bugs at \url{https://github.com/r-lib/covr/issues}
}

}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com}

Other contributors:
\itemize{
  \item Willem Ligtenberg [contributor]
  \item Kirill Müller [contributor]
  \item Henrik Bengtsson [contributor]
  \item Steve Peak [contributor]
  \item Kirill Sevastyanenko [contributor]
  \item Jon Clayden [contributor]
  \item Robert Flight [contributor]
  \item Eric Brown [contributor]
  \item Brodie Gaslam [contributor]
  \item Will Beasley [contributor]
  \item Robert Krzyzanowski [contributor]
  \item Markus Wamser [contributor]
  \item Karl Forner [contributor]
  \item Gergely Daróczi [contributor]
  \item Jouni Helske [contributor]
  \item Kun Ren [contributor]
  \item Jeroen Ooms [contributor]
  \item Ken Williams [contributor]
  \item Chris Campbell [contributor]
  \item David Hugh-Jones [contributor]
  \item Qin Wang [contributor]
  \item Ivan Sagalaev (highlight.js library) [contributor, copyright holder]
  \item Mark Otto (Bootstrap library) [contributor]
  \item Jacob Thornton (Bootstrap library) [contributor]
  \item  Bootstrap contributors (Bootstrap library) [contributor]
  \item  Twitter, Inc (Bootstrap library) [copyright holder]
}

}
