% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{covr.record_tests}
\alias{covr.record_tests}
\title{Record Test Traces During Coverage Execution}
\description{
By setting \code{options(covr.record_tests = TRUE)}, the result of covr coverage
collection functions will include additional data pertaining to the tests
which are executed and an index of which tests, at what stack depth, trigger
the execution of each trace.
}
\section{Additional fields}{


Within the \code{covr} result, you can explore this information in two places:

\itemize{
\item \code{attr(,"tests")}: A list of call stacks, which results in target code
execution.

\item \verb{$<srcref>$tests}: For each srcref count in the coverage object, a
\verb{$tests} field is now included which contains a matrix with two columns,
"test" and "depth" which specify the test number (corresponding to the
index of the test in \code{attr(,"tests")} and the stack depth into the target
code where the trace was executed.
}
}

\examples{
fcode <- '
f <- function(x) {
  if (x)
    TRUE
  else
    FALSE
}'

options(covr.record_tests = TRUE)
cov <- code_coverage(fcode, "f(TRUE)")

# extract executed tests traces
attr(cov, "tests")
# $`/tmp/test.R:1:1:1:7:1:7:1:1`
# $`/tmp/test.R:1:1:1:7:1:7:1:1`[[1]]
# f(TRUE)

# extract test itemization per trace
cov[[3]][c("srcref", "tests")]
# $srcref
# TRUE
#
# $tests
#      test depth
# [1,]    1     1

}
