% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_name.R
\name{display_name}
\alias{display_name}
\title{Retrieve the display name from an object}
\usage{
display_name(x, ...)
}
\arguments{
\item{x}{object from which to retrieve the display name}

\item{...}{additional arguments passed to methods}
}
\description{
Retrieve the display name from an object
}

