% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covr.R
\name{environment_coverage}
\alias{environment_coverage}
\title{calculate the coverage on an environment after evaluating some expressions.}
\usage{
environment_coverage(env, ..., enc = parent.frame())
}
\arguments{
\item{env}{the environment to take function definitions from}

\item{...}{one or more expressions to be evaluated.}

\item{enc}{The enclosing environment from which the expressions should be
evaluated}
}
\description{
This function uses non_standard evaluation so is best used in interactive
sessions.
}

