% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr.R
\name{environment_coverage}
\alias{environment_coverage}
\title{Calculate coverage of an environment}
\usage{
environment_coverage(
  env = parent.frame(),
  test_files,
  line_exclusions = NULL,
  function_exclusions = NULL
)
}
\arguments{
\item{env}{The environment to be instrumented.}

\item{test_files}{Character vector of test files with code to test the
functions}

\item{line_exclusions}{a named list of files with the lines to exclude from
each file.}

\item{function_exclusions}{a vector of regular expressions matching function
names to exclude. Example \verb{print\\\\\\.} to match print methods.}
}
\description{
Calculate coverage of an environment
}
\examples{
my_env <-  new.env()
my_env$add <- function(x, y) { x + y }
cat("add(1, 2) == 3", file="add_test.R")
environment_coverage(my_env, test_files = "add_test.R")
}
