% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{gitlab}
\alias{gitlab}
\title{Run covr on package and create report for GitLab}
\usage{
gitlab(..., coverage = NULL, file = "public/coverage.html", quiet = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{\link[=package_coverage]{package_coverage()}}}

\item{coverage}{an existing coverage object to submit, if \code{NULL},
\code{\link[=package_coverage]{package_coverage()}} will be called with the arguments from
\code{...}}

\item{file}{The report filename.}

\item{quiet}{if \code{FALSE}, print the coverage before submission.}
}
\description{
Utilize internal GitLab static pages to publish package coverage.
Creates local covr report in a package subdirectory.
Uses the \href{https://docs.gitlab.com/user/project/pages/}{pages}
GitLab job to publish the report.
}
