% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_tests.R
\name{is_current_test_finished}
\alias{is_current_test_finished}
\title{Returns TRUE if we've moved on from test reflected in .current_test}
\usage{
is_current_test_finished()
}
\description{
Returns TRUE if we've moved on from test reflected in .current_test
}
