% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{new_counter}
\alias{new_counter}
\title{initialize a new counter}
\usage{
new_counter(src_ref, enclosing_funs)
}
\arguments{
\item{src_ref}{a \code{\link[base]{srcref}}}

\item{enclosing_funs}{the functions that this srcref is contained in.}
}
\description{
initialize a new counter
}

