% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{percent_coverage}
\alias{percent_coverage}
\title{Provide percent coverage of package}
\usage{
percent_coverage(x, ...)
}
\arguments{
\item{x}{the coverage object returned from \code{\link{package_coverage}}}

\item{...}{additional arguments passed to \code{\link{tally_coverage}}}
}
\description{
Print the total percent coverage
}

