% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{percent_coverage}
\alias{percent_coverage}
\title{Provide percent coverage of package}
\usage{
percent_coverage(x, ...)
}
\arguments{
\item{x}{the coverage object returned from \code{\link{package_coverage}}}

\item{...}{additional arguments passed to \code{\link{tally_coverage}}}
}
\value{
The total percentage as a \code{numeric(1)}.
}
\description{
Calculate the total percent coverage from a coverage result object.
}
