% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{print.coverage}
\alias{print.coverage}
\title{Print a coverage object}
\usage{
\method{print}{coverage}(x, group = c("filename", "functions"), by = "line", ...)
}
\arguments{
\item{x}{the coverage object to be printed}

\item{group}{whether to group coverage by filename or function}

\item{by}{whether to count coverage by line or expression}

\item{...}{additional arguments ignored}
}
\value{
The coverage object (invisibly).
}
\description{
Print a coverage object
}
