% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{shine}
\title{Display covr results using a standalone report}
\usage{
report(x, ...)
}
\arguments{
\item{x}{a coverage dataset}

\item{browse}{whether to open a browser to veiew the report.}
}
\description{
Display covr results using a standalone report
}
\examples{
\dontrun{
x <- package_coverage()
report(x)
}
}

