% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{report.coverage}
\alias{shine}
\title{Display covr results using a standalone report}
\usage{
report(x, ...)

shine(x, ...)

\method{report}{coverage}(x, file = file.path(tempdir(),
  paste0(get_package_name(x), "-report.html")), browse = interactive(), ...)
}
\arguments{
\item{x}{a coverage dataset}

\item{...}{Additional arguments passed to methods}

\item{file}{The report filename.}

\item{browse}{whether to open a browser to view the report.}
}
\description{
Display covr results using a standalone report
}
\examples{
\dontrun{
x <- package_coverage()
report(x)
}
}

