% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Display covr results using a standalone report}
\usage{
report(x = package_coverage(), file = file.path(tempdir(),
  paste0(get_package_name(x), "-report.html")), browse = interactive())
}
\arguments{
\item{x}{a coverage dataset, defaults to running \code{package_coverage()}.}

\item{file}{The report filename.}

\item{browse}{whether to open a browser to view the report.}
}
\description{
Display covr results using a standalone report
}
\examples{
\dontrun{
x <- package_coverage()
report(x)
}
}
