% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny.R
\name{shine}
\alias{shine}
\title{Display covr results using a shiny app}
\usage{
shine(x)
}
\arguments{
\item{x}{a coverage dataset
\dontrun{
x <- package_coverage()
shine(x)
}}
}
\description{
The shiny app is designed to provide local information to coverage
information similar to the coveralls.io website.  However it does not and
will not track coverage over time.
}

