% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subprocess.R
\name{subprocess}
\alias{subprocess}
\title{Run R commands in a R subprocess}
\usage{
subprocess(..., calling_env = parent.frame(), global_env = .GlobalEnv,
  quiet = TRUE, clean_output = TRUE)
}
\arguments{
\item{...}{expressions to call in the subprocess}

\item{calling_env}{the environment of the calling function.}

\item{global_env}{the global environment to load in the subprocess.}

\item{quiet}{whether to echo the R command run.}

\item{clean_output}{whether to clean the Rout files generated from the subprocess.}
}
\description{
this function runs a R subprocess with expressions given in ..., trying to
replicate as much of the calling environment as possible, then returning the
results to the calling environment.
}

