% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_calls.R
\name{trace_calls}
\alias{trace_calls}
\title{trace each call with a srcref attribute}
\usage{
trace_calls(x, parent_functions = NULL, parent_ref = NULL)
}
\arguments{
\item{x}{the call}

\item{parent_functions}{the functions which this call is a child of.}

\item{parent_ref}{argument used to set the srcref of the current call when recursing}
}
\value{
a modified expression with count calls inserted before each previous
call.
}
\description{
This function calls itself recursively so it can properly traverse the AST.
}
\seealso{
http://adv-r.had.co.nz/Expressions.html
}
\keyword{internal}
