% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{zero_coverage}
\alias{zero_coverage}
\title{Provide locations of zero coverage}
\usage{
zero_coverage(x, ...)
}
\arguments{
\item{x}{a coverage object returned \code{\link{package_coverage}}}

\item{...}{additional arguments passed to
\code{\link{tally_coverage}}}
}
\description{
When examining the test coverage of a package, it is useful to know if there are
any locations where there is \bold{0} test coverage.
}

